notes on the firefox bug
basically we have to set focus ourselves. This means when we start
managing a window, and when an XCrossingEvent occurs (which we don't
handle).

On Manage/Enter, we set focus. on Leave we set focus to root.
See event.c and focus.c in dwm for more details.

darcs-hash:20070309162510-9c5c1-142ded89f618d4234b7ad79265f63e9f5fd9f42f