Focus follows mouse.
This change makes the window under the mouse pointer the focused window.  This
isn't quite what we want, but it is a step in the right direction.  The next
step is to somehow inhibit the CrossingEvents generated during workspace and
layout switches.

darcs-hash:20070326124725-a5988-825770ffc28620d16b7aba0541e3abff29b46680