make workspace tag not need to be a Num.
This change also removes the barely used 'size' field, and replaces
it with a tagMember predicate.  The idea is to move towards the ability
to make the workspace tag be a String, which by default might be "1".."9",
but could also be customized to be something meaningful to the user.

darcs-hash:20070614140709-72aca-dbc1ce07bf964e4cff69deef95b453a0051ef782