Add generated manpage and html manpage to the repo.
Ignore-this: e588bc3ad31ea51ceaa9316c75ab6633

The intention of adding these files to the data-files
is so that they get included in the upload to hackage:
people might like manpage but not have to install pandoc.

It's not really clear that this is the best solution.

darcs-hash:20121108231139-1499c-a88dc1c44371ebb92b1254b07b530c3682176213